requirements.txt
azure-cognitiveservices-speech==1.23.0
flask==2.2.2
wave==0.0.2
soundfile==0.11.0
librosa==0.9.2